Rods + Contacts Working

+++ GDPHYSX_Creencia_Villegas/PhysicsWorld.cpp
@@ -14,7 +14,6 @@ void PhysicsWorld::update(float time)
 			particles[i]->update(time);
 		}
 
-
 		generateContacts();
 
 		if (contacts.size() > 0)
@@ -23,6 +22,14 @@ void PhysicsWorld::update(float time)
 			contactResolver.resolveContacts(contacts, time);
 		}
 	}
+
+	// Update particle positions for drawing
+	for (int i = 0; i < particles.size(); i++) 
+	{
+		particleTrans[i] = glm::mat4(1.0f);
+		particleTrans[i] = glm::scale(particleTrans[i], glm::vec3(0.5f, 0.5f, 0.5f));
+		particleTrans[i] = glm::translate(particleTrans[i], glm::vec3(particles[i]->getVec3Pos()));
+	}
 }
 
 void PhysicsWorld::addParticle(int pType, MyVector currParticlePos)
@@ -30,8 +37,9 @@ void PhysicsWorld::addParticle(int pType, MyVector currParticlePos)
 	MyParticle* newParticle = new MyParticle(pType);
 
 	glm::mat4 trans = glm::mat4(1.0f); // identity
-	trans = glm::translate(trans, glm::vec3(currParticlePos.x, currParticlePos.y , currParticlePos.z));
-	trans = glm::scale(trans, glm::vec3(0.5f, 0.5f, 0.5f));
+	if (pType == 0) trans = glm::scale(trans, glm::vec3(0.1f, 0.1f, 0.1f));
+	else trans = glm::scale(trans, glm::vec3(0.5f, 0.5f, 0.5f));
+	trans = glm::translate(trans, glm::vec3(currParticlePos.x, currParticlePos.y, currParticlePos.z));
 	particleTrans.push_back(trans);
 
 	glm::mat4 normalTrans;
@@ -65,18 +73,11 @@ void PhysicsWorld::deleteParticle(int index)
 	particles.erase(particles.begin() + index);
 }
 
-void PhysicsWorld::clearParticles()
-{
-	particleTrans.clear();
-	normalTransArray.clear();
-	particles.clear();
-}
-
 void PhysicsWorld::updateParticleList()
 {
 	for (int i = 0; i < particles.size(); i++)
 	{
-		if (particles[i]->isDestroyed)
+		if (particles[i]->isDestroyed && particles[i]->type != 0)
 		{
 			registry.remove(particles[i], &Gravity);
 			deleteParticle(i);
@@ -88,9 +89,9 @@ void PhysicsWorld::generateContacts()
 {
 	contacts.clear();
 	getOverlaps();
-	for (list<ParticleLink*>::iterator i = links.begin(); i != links.end(); i++)
+	for (int i = 0; i < links.size(); i++)
 	{
-		ParticleContact* contact = (*i)->getContact();
+		ParticleContact* contact = links[i]->getContact();
 		if (contact != nullptr)
 		{
 			contacts.push_back(contact);
